Refactor JSROOT read function for better readability
